修复bug 获取真实id

+++ src/main/java/com/gistone/demomybatis/web/IndexController.java
@@ -23,11 +23,11 @@ public class IndexController {
     public InsertResult in10() {
         StopWatch stopWatch = new StopWatch();
         stopWatch.start();
-        int insert = studentService.insert(Student.randOne());
+        long insert = studentService.insert(Student.randOne());
         stopWatch.stop();
         return new InsertResult(
                 stopWatch.getTotalTimeMillis(),
-                Long.valueOf(insert)
+                insert
         );
     }
 
@@ -38,8 +38,8 @@ public class IndexController {
         final int total = 10;
         InsertResult insertResult = new InsertResult();
         for (int i = 0; i < total; i++) {
-            int insert = studentService.insert(Student.randOne());
-            insertResult.add((long) insert);
+            long insert = studentService.insert(Student.randOne());
+            insertResult.add(insert);
         }
         stopWatch.stop();
         insertResult.setInsertTime(stopWatch.getTotalTimeMillis());
